requirements.txt
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
requests==2.31.0
python-dotenv==1.0.0
urllib3==2.0.7
instagrapi==2.1.3
selenium==4.18.1
webdriver-manager==4.0.1
